Migration: Refactoring of Utilities.
