+++ ts/Core/Axis/Axis.ts
@@ -35,7 +35,10 @@ import Color from '../Color/Color.js';
 import H from '../Globals.js';
 import palette from '../Color/Palette.js';
 import O from '../Options.js';
-const { defaultOptions } = O;
+const {
+    defaultOptions,
+    format
+} = O;
 import Tick from './Tick.js';
 import U from '../Utilities.js';
 const {
@@ -50,7 +53,6 @@ const {
     error,
     extend,
     fireEvent,
-    format,
     getMagnitude,
     isArray,
     isFunction,
