+++ ts/Core/Chart/StockChart.ts
@@ -26,6 +26,12 @@ import type SVGElement from '../Renderer/SVG/SVGElement';
 import type SVGPath from '../Renderer/SVG/SVGPath';
 import Axis from '../Axis/Axis.js';
 import Chart from '../Chart/Chart.js';
+import O from '../../Core/Options.js';
+const {
+    format,
+    getOptions,
+    setOptions
+} = O;
 import palette from '../../Core/Color/Palette.js';
 import Point from '../Series/Point.js';
 const {
@@ -50,8 +56,6 @@ const {
     defined,
     extend,
     find,
-    format,
-    getOptions,
     isNumber,
     isString,
     merge,
