+++ ts/Extensions/ParallelCoordinates.ts
@@ -19,6 +19,11 @@ import type SeriesOptions from '../Core/Series/SeriesOptions';
 import Axis from '../Core/Axis/Axis.js';
 import Chart from '../Core/Chart/Chart.js';
 import H from '../Core/Globals.js';
+import O from '../Core/Options.js';
+const {
+    format,
+    setOptions
+} = O;
 import Series from '../Core/Series/Series.js';
 import U from '../Core/Utilities.js';
 const {
@@ -28,10 +33,8 @@ const {
     defined,
     erase,
     extend,
-    format,
     merge,
     pick,
-    setOptions,
     splat,
     wrap
 } = U;
