+++ ts/Extensions/Themes/GridLight.ts
@@ -15,11 +15,10 @@
 /* global document */
 
 import Highcharts from '../../Core/Globals.js';
+import O from '../../Core/Options.js';
+const { setOptions } = O;
 import U from '../../Core/Utilities.js';
-const {
-    createElement,
-    setOptions
-} = U;
+const { createElement } = U;
 
 // Load the fonts
 createElement('link', {
