+++ ts/Extensions/Themes/SandSignika.ts
@@ -15,11 +15,12 @@
 /* global document */
 
 import Highcharts from '../../Core/Globals.js';
+import O from '../../Core/Options.js';
+const { setOptions } = O;
 import U from '../../Core/Utilities.js';
 const {
     addEvent,
-    createElement,
-    setOptions
+    createElement
 } = U;
 
 // Load the fonts
