+++ ts/Series/Bubble/BubbleLegend.ts
@@ -27,6 +27,8 @@ const { parse: color } = Color;
 import H from '../../Core/Globals.js';
 const { noop } = H;
 import Legend from '../../Core/Legend.js';
+import O from '../../Core/Options.js';
+const { setOptions } = O;
 import palette from '../../Core/Color/Palette.js';
 import Series from '../../Core/Series/Series.js';
 import U from '../../Core/Utilities.js';
@@ -39,7 +41,6 @@ const {
     merge,
     objectEach,
     pick,
-    setOptions,
     stableSort,
     wrap
 } = U;
@@ -893,7 +894,7 @@ class BubbleLegend {
             format = (options.labels as any).format;
         const { numberFormatter } = this.chart;
 
-        return format ? U.format(format, range) :
+        return format ? O.format(format, range) :
             formatter ? formatter.call(range) :
                 numberFormatter(range.value, 1);
     }
